+++ source/App/source/WebApp/Extensions/DependencyInjection/HealthChecksExtensions.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.App.Common.Diagnostics.HealthChecks;
+using Energinet.DataHub.Core.App.Common.Extensions.Builder;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Energinet.DataHub.Core.App.WebApp.Extensions.DependencyInjection;
